ハンバーガーメニュー内のリンククリック時、メニューを閉じる修正 トップページアクセス時、sessionの状況によってアクセス先振り分け
